pyproject.toml
[tool.poetry]
name = "messageapi"
version = "0.1.0"
description = ""
authors = ["glen15 <glen910115@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.0"
djangorestframework = "3.13.0"
flake8 = "^6.0.0"
black = "^22.10.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
